requirements.txt
# ===========================================
# Python Script 依赖管理
# 最后更新: 2026-04-23
# ===========================================

# -------------------------------------------
# 数据处理
# -------------------------------------------
pandas==2.0.1
numpy>=1.24.0
openpyxl==3.0.10
xlrd==2.0.1
xlwt==1.3.0

# -------------------------------------------
# 数据库
# -------------------------------------------
pymysql==1.1.0
SQLAlchemy==2.0.25
impala==0.2

# -------------------------------------------
# 中间件
# -------------------------------------------
kafka-python==2.0.2
redis==4.5.5

# -------------------------------------------
# Web 框架
# -------------------------------------------
flask==2.3.0

# -------------------------------------------
# PDF 处理
# -------------------------------------------
PyMuPDF==1.22.5
PyPDF2==3.0.1
PyPDF4==1.27.0
pdf2docx==0.5.6
pdfminer==20191125
pikepdf>=8.0.0

# -------------------------------------------
# Office 文档处理
# -------------------------------------------
python-docx>=0.8.11
python-pptx>=0.6.21

# -------------------------------------------
# 图片处理
# -------------------------------------------
Pillow>=9.5.0
rembg>=2.0.0

# -------------------------------------------
# 网络请求与爬虫
# -------------------------------------------
requests>=2.28.0
beautifulsoup4>=4.12.0
selenium>=4.8.0

# -------------------------------------------
# NLP 与 AI
# -------------------------------------------
nltk==3.6.2
textblob>=0.17.1
openai>=1.0.0

# -------------------------------------------
# 工具类
# -------------------------------------------
faker>=18.0.0
paramiko>=3.0.0
python-hosts>=1.0.5
pyshorteners>=1.0.1
pyttsx3>=2.90
pywebview>=4.0
pywhatkit>=5.4
avro-python3>=1.10.0
pycryptodome>=3.18.0

# -------------------------------------------
# 开发工具
# -------------------------------------------
python-dotenv>=1.0.0